Always render system errors on exit regardless of fixer status
Previously, system errors were only rendered if running in `no fix`
mode. When running with `--fix` then the ConsoleOutputFormatter
only calls `printAfterFixerStatus()` and this did not display
system errors.

This meant that the command could fail silently with a nonzero exit
code (because the ExitCodeResolver exits 1 if there were any system
errors, regardless of mode).

It looks like this was unintentional -
https://github.com/symplify/symplify/commit/d08859d3808f69d1b8502ece5d7a86a8077c5058
added the code to split system errors to a separate stream but
only added code to render them in `printNoFixerStatus`

I think this is related to / masking some of the various caching
bugs that have been reported previously e.g. #29 which mentions
the process failing silently without any error message. In our
case the system errors ECS encountered were also due to cache write
failures.