refactor(formatters): improved discoverability of formatters & their properties
Whether or not a formatter supports progress bars is now encapsulated
in the formatter class itself, as opposed to being scattered in a couple places
and relying on maintainers remembering.

Both this progress bar support and the formatter names are also
implemented as static methods now, whose inclusion can be
promised by their shared interface.

The console output retains its `::NAME` constant because it needs to be used
in the default parameters of a configuration constructor (being a default
setting).

Finally, there remains one hard-coded reference to the `json` output format,
but only over in the `ListCheckersCommand`. I've left this alone since this
is not the same type of formatter. Don't want to step on any toes. :)