Update Composer Linting Scripts (#211)
* build: add composer script aliases to simplify builds

I added a separate `fix-rector` along the lines of `fix-cs` to respect the
original author's preferred approach. The aliases just make it easy for other
contributors to run all the tools in the correct order without having to
look it up.

* build: increased memory limit for PhpStan

The default of 128M was causing consistent crashes for me on Github Codespaces.

Since Rector uses PhpStan in the background, I also upped it's limit. Though
this really shouldn't matter, since Rector can use the existing cache and
needs less memory anyway.