Add Gitlab / Code Climate Output Format Support (#210)
* refactor(formatters): improved discoverability of formatters & their properties

Whether or not a formatter supports progress bars is now encapsulated
in the formatter class itself, as opposed to being scattered in a couple places
and relying on maintainers remembering.

Both this progress bar support and the formatter names are also
implemented as static methods now, whose inclusion can be
promised by their shared interface.

The console output retains its `::NAME` constant because it needs to be used
in the default parameters of a configuration constructor (being a default
setting).

Finally, there remains one hard-coded reference to the `json` output format,
but only over in the `ListCheckersCommand`. I've left this alone since this
is not the same type of formatter. Don't want to step on any toes. :)

* feat(formatters): added a Gitlab/Code Climate output formatter

* docs(formatters): added a README section on output formats for clarity

I, personally, didn't realize this tool supported output formatting at first,
and the existing issues seemed to have been closed without mentioning
the addition of them. I'd just like to make it more obvious for new users!

* refactor(formatters): ran automated fixers on changed files

* fix(formatters): set gitlab issue severity to 'minor'

My initial line of thinking about setting to 'blocker' was flawed. Not everyone
uses this tooling the same way and they may want it to show up in the MR
widget/diffs without _actually_ blocking. Since ECS focuses on code style
standards instead of correctness issues, this seems an appropriate level.

* refactor(formatters): create separate method to generate report string

This just makes testing easier, as seen in the existing test classes.

* chore: stop linting test fixtures, which may intentionally contain errors

* test(formatters): initial testing for the Gitlab output formatting

This is by no means comprehensive, but it at least confirms it respects
configurations as expected, formatting is right for trivial cases, and that
fingerprints are deterministic.

* test(formatters): minor test improvement for Gitlab formatter

- Uncommenting an attribute I disabled for debugging.
- Added an important clarifying comment.

* test(formatters): fix incorrect paths (and fingerprints) for Gitlab formatter

These should have been relative, but they were absolute. There's no correctness
issue in the formatter itself, it just repeats the paths it's given. But test
cases should generally be correct regardless.

---------

Co-authored-by: Tomas Votruba <tomas.vot@gmail.com>