fix(caching): improve reliability of configuration caching
The output of `getBindings()` is largely just a list of the class names
registered in the container, but it doesn't contain information on how
those classes are actually configured. As such, changing `ecs.php` in a way
that updates sniff / fixer rules would not invalidate the cache, meaning
those new settings would not be checked against any unchanged files.

My approach here still isn't perfectly reliable, but I think it's a usable
middle ground. We still report all the registered classes, but
now I explicitly pull out the configuration data for sniffers and fixers
to add to the serialized data. Then I also hash out the settings used for our
sniffers, to support the upcoming PHPCS rulesets / standards support.

The ECSConfig previously constructed as-needed, is now provided via DI.
This is required for the upcoming feature, but has the added benefits of making
sure the cache hash uses the hash for the exact same config that will be used
for sniffing *and* improving the performance of sniffing
(since every file is currently creating a config, triggering filesystem ops).

Closes #30