feat(phpcs): added support for standard imports and sniffer warnings.
This mostly involved redefining the entire code path on how we determine
when sniffs are skipped over to reflect more what PHPCS does in the parent file
WRT to sniff severity at a code-level granularity (as opposed to class-level).
I, unfortunately, didn't see a great way to re-use their work directly via
`parent::` method invocation, we just deviate too much.

All the additional `add*` methods are necessary on a sniff-by-sniff basis.
*Technically* this is a fix for an existing bug, where some sniffs would be
unable to report issues to ECS correctly, but they're not really widely used
or only applied to warnings - which we universally ignored before - so no
need to make a ticket about it.

As you can see in the prior changes to `README.md`, standard / ruleset imports
are very much a partially-supported feature. There's a lot I just can't
replicate without a significant undertaking doing refactors. And I'm not
sure it would fit the ECS founders' intention of his tooling, either -
I certainly want to respect that.

I left in the previously defined warning escalations, which completely override
ruleset / standard severity from PHPCS configs, though I'm not quite sure
why we do that or if my integration is exactly how we want to continue
integrating it...

Closes #128