feat: add initial support for `.editorconfig`
The approach I've gone for puts `.editorconfig` settings above those settings in
the `ecs.php` config itself, to prioritize convenience for those people using
pre-built sets. As many setttings as I could easily implement have been
supported, including one proposed (but not standardized) setting I like to
use.

This commit includes tests, documentation, and dogfooding.

Tests only currently cover the parsing of `.editorconfig`, but do not test
how all permutations affect configuration. Doing so would require advanced
alias mocks of the `EditorConfigFactory` and it's `load` method, which I wasn't
able to get working in an initial attempt. That, or a refactor of
`ECSConfigBuilder` to allow injecting the factory, but I didn't want to mess
with  any of the exposed public API.

Closes #217